Added scope maxDescendantChildren.
With scope, all descendants of the parent Element 
of the Processing Instruction are considered. Only
 child Elements of those secondary, descendant 
Elements are counted. The maximum count made for 
any	of these descendant Elements is stored in the 
parent Element of the Processing Instruction. This
 is useful for establishing the dimension size 
required, when storing static XML data in 
multi-dimensional, runtime data structures.